feat(flutter): reanimated project. added ios support

+++ lib/screens/menu/account.dart
@@ -8,37 +8,36 @@ import 'package:flutter/material.dart';
 
 import 'package:coworking/domain/services/sign_in.dart';
 
-//Максимальная и минимальная длина имени
+// Максимальная и минимальная длина имени
 const int userNameMin = 1;
 const int userNameMax = 100;
 
 class AccountScreen extends StatelessWidget {
+  AccountScreen({super.key});
   final GlobalKey<DisplayNameFormState> formKey = GlobalKey();
-
-  AccountScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       resizeToAvoidBottomInset: false,
       appBar: AppBar(
-        title: const Text("Настройки аккаунта"),
+        title: const Text('Настройки аккаунта'),
         actions: <Widget>[
           PopupMenuButton(
-            tooltip: "Help",
+            tooltip: 'Help',
             icon: const Icon(
               Icons.help,
               color: Colors.black,
             ),
             itemBuilder: (BuildContext context) => <PopupMenuEntry>[
               const PopupMenuItem(
-                child: Text("\nЭто страница Вашего профиля.\n"
-                    "\nВы можете поменять свое имя, посмотреть количество "
-                    "посещенных Вами мест, написанных отзывов "
-                    "и выйти либо удалить аккаунт.\n"),
+                child: Text('\nЭто страница Вашего профиля.\n'
+                    '\nВы можете поменять свое имя, посмотреть количество '
+                    'посещенных Вами мест, написанных отзывов '
+                    'и выйти либо удалить аккаунт.\n'),
               ),
             ],
-          )
+          ),
         ],
       ),
       body: GestureDetector(
@@ -69,56 +68,60 @@ class AccountScreen extends StatelessWidget {
               DisplayNameForm(key: formKey),
               const SizedBox(height: 32.0),
               Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  children: <Widget>[
-                    Column(children: [
+                mainAxisAlignment: MainAxisAlignment.spaceAround,
+                children: <Widget>[
+                  Column(
+                    children: [
                       FutureBuilder(
                         future: DatabasePin.fetchFavouritePinsAmount(),
                         builder: (context, snapshot) {
                           if (snapshot.hasData) {
-                            return Text(snapshot.data!.toString(),
-                                textScaleFactor: 2.0);
-                          } else {
-                            return const CircularProgressIndicator();
-                          }
-                        },
-                      ),
-                      const Text("Посещенные места"),
-                    ]),
-                    Column(children: [
-                      FutureBuilder(
-                        future: DatabaseReview.fetchReviewsOfUserAmount(
-                            Account.currentAccount!),
-                        builder: (context, snapshot) {
-                          if (snapshot.hasData) {
                             return Text(
                               snapshot.data!.toString(),
-                              textScaleFactor: 2.0,
+                              textScaler: const TextScaler.linear(2.0),
                             );
                           } else {
                             return const CircularProgressIndicator();
                           }
                         },
                       ),
-                      const Text("Отзывов написано"),
-                    ]),
-                  ]),
+                      const Text('Посещенные места'),
+                    ],
+                  ),
+                  Column(
+                    children: [
+                      FutureBuilder(
+                        future: DatabaseReview.fetchReviewsOfUserAmount(Account.currentAccount!),
+                        builder: (context, snapshot) {
+                          if (snapshot.hasData) {
+                            return Text(
+                              snapshot.data!.toString(),
+                              textScaler: const TextScaler.linear(2.0),
+                            );
+                          } else {
+                            return const CircularProgressIndicator();
+                          }
+                        },
+                      ),
+                      const Text('Отзывов написано'),
+                    ],
+                  ),
+                ],
+              ),
               const Spacer(),
               ElevatedButton(
                 onPressed: () => signOut(context),
                 style: ButtonStyle(
-                  backgroundColor:
-                      MaterialStateProperty.all<Color>(Colors.orange),
+                  backgroundColor: WidgetStateProperty.all<Color>(Colors.orange),
                 ),
-                child: const Text("Выйти из профиля"),
+                child: const Text('Выйти из профиля'),
               ),
               ElevatedButton(
                 onPressed: () => handleDeleteButton(context),
                 style: ButtonStyle(
-                  backgroundColor: MaterialStateProperty.all<Color>(
-                      Theme.of(context).errorColor),
+                  backgroundColor: WidgetStateProperty.all<Color>(Colors.red),
                 ),
-                child: const Text("Удалить профиль"),
+                child: const Text('Удалить профиль'),
               ),
             ],
           ),
@@ -132,30 +135,29 @@ class AccountScreen extends StatelessWidget {
       context: context,
       barrierDismissible: false,
       builder: (BuildContext context) => AlertDialog(
-        title: const Text("Вы уверены?"),
-        content: const Text(
-            "Все ваши встречи будут удалены, но пины и отзывы останутся."),
+        title: const Text('Вы уверены?'),
+        content: const Text('Все ваши встречи будут удалены, но пины и отзывы останутся.'),
         actions: <Widget>[
           TextButton(
             onPressed: () => Navigator.of(context).pop(false),
-            child: const Text("Отмена"),
+            child: const Text('Отмена'),
           ),
           TextButton(
             onPressed: () => Navigator.of(context).pop(true),
-            child: const Text("Удалить", style: TextStyle(color: Colors.red)),
-          )
+            child: const Text('Удалить', style: TextStyle(color: Colors.red)),
+          ),
         ],
       ),
     );
 
-    if (confirmed != null) {
+    if (confirmed != null && context.mounted) {
       deleteAccount(context);
     }
   }
 }
 
 class DisplayNameForm extends StatefulWidget {
-  const DisplayNameForm({Key? key}) : super(key: key);
+  const DisplayNameForm({super.key});
 
   @override
   State<DisplayNameForm> createState() => DisplayNameFormState();
@@ -171,22 +173,22 @@ class DisplayNameFormState extends State<DisplayNameForm> {
 
   @override
   void initState() {
-    controller = TextEditingController(
-        text: FirebaseAuth.instance.currentUser!.displayName);
+    controller = TextEditingController(text: FirebaseAuth.instance.currentUser!.displayName);
 
     super.initState();
   }
 
-  void submitValue(value) {
+  void submitValue(String? value) {
+    if (value == null) return;
     String? oldDisplayName = FirebaseAuth.instance.currentUser!.displayName;
     DatabaseAccount.updateUsername(value);
 
     setState(() => pending = false);
 
     SnackBar snackBar = SnackBar(
-      content: const Text("Ваше имя было изменено"),
+      content: const Text('Ваше имя было изменено'),
       action: SnackBarAction(
-        label: "Отменить",
+        label: 'Отменить',
         onPressed: () {
           DatabaseAccount.updateUsername(oldDisplayName!);
           setState(() {
@@ -208,9 +210,10 @@ class DisplayNameFormState extends State<DisplayNameForm> {
       focusNode: formFocus,
       decoration: InputDecoration(
         icon: const Icon(Icons.person),
-        labelText: "Ваше имя",
-        hintText: "Напишите свое имя",
+        labelText: 'Ваше имя',
+        hintText: 'Напишите свое имя',
         suffixIcon: Visibility(
+          visible: pending,
           child: IconButton(
             icon: const Icon(Icons.save),
             onPressed: () {
@@ -218,7 +221,6 @@ class DisplayNameFormState extends State<DisplayNameForm> {
               key.currentState!.save();
             },
           ),
-          visible: pending,
         ),
       ),
       onChanged: (_) {
@@ -234,15 +236,13 @@ class DisplayNameFormState extends State<DisplayNameForm> {
     RegExp alphaNumRegEx = RegExp(r'^[a-zA-Z0-9]+$');
 
     if (!alphaNumRegEx.hasMatch(value!) && value.isNotEmpty) {
-      return "Ваше имя может состоять только из букв";
+      return 'Ваше имя может состоять только из букв';
     }
     if (value.length > userNameMax) {
-      return "Ваше имя слишком длинное, максимальная длинна " +
-          userNameMax.toString();
+      return 'Ваше имя слишком длинное, максимальная длинна $userNameMax';
     }
     if (value.length < userNameMin) {
-      return "Ваше имя слишком короткое, минимальная длинна " +
-          userNameMin.toString();
+      return 'Ваше имя слишком короткое, минимальная длинна $userNameMin';
     }
     return null;
   }
@@ -254,14 +254,14 @@ void deleteAccount(BuildContext context) async {
   DatabaseAccount.deleteUser(Account.currentAccount!);
   await currentUser!.delete();
   SignIn().signOutGoogle();
-  Navigator.of(context)
-      .popUntil(ModalRoute.withName(MainNavigationRouteNames.auth));
+  if (context.mounted) {
+    Navigator.of(context).popUntil(ModalRoute.withName(MainNavigationRouteNames.auth));
+  }
 }
 
 void signOut(BuildContext context) {
   FirebaseAuth.instance.signOut();
   SignIn().signOutGoogle();
   Account.currentAccount = null;
-  Navigator.of(context)
-      .popUntil(ModalRoute.withName(MainNavigationRouteNames.auth));
+  Navigator.of(context).popUntil(ModalRoute.withName(MainNavigationRouteNames.auth));
 }
